+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - proxy: add the location of a caller to panics ([#75]).
 - telemeter: make `elfo_metrics_usage_bytes` more accurate.
+- context: decrement a budget only when an envelope has been actually received.
 
 [#77]: https://github.com/elfo-rs/elfo/pull/77
 [#76]: https://github.com/elfo-rs/elfo/pull/76